# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"passwordHash": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"phone": {
						"type": "number",
						"format": "double"
					},
					"admin": {
						"type": "boolean"
					},
					"image": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"firstName",
					"lastName",
					"email",
					"passwordHash",
					"address",
					"phone",
					"admin",
					"image",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserPayload": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"passwordHash": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"phone": {
						"type": "number",
						"format": "double"
					},
					"admin": {
						"type": "boolean"
					},
					"data": {
						"type": "string",
						"format": "byte"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"email",
					"passwordHash",
					"address",
					"phone",
					"admin",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Boolean": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"Cart": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"cartItems": {
						"items": {
							"$ref": "#/components/schemas/CartItems"
						},
						"type": "array"
					},
					"finished": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"user",
					"userId",
					"cartItems",
					"finished",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Category": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"image",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Promo": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"discount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"image",
					"discount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Item": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"$ref": "#/components/schemas/Category"
					},
					"categoryId": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"promo": {
						"$ref": "#/components/schemas/Promo"
					},
					"promoId": {
						"type": "number",
						"format": "double"
					},
					"image": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"category",
					"categoryId",
					"name",
					"description",
					"price",
					"promo",
					"promoId",
					"image",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CartItems": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"cart": {
						"$ref": "#/components/schemas/Cart"
					},
					"cartId": {
						"type": "number",
						"format": "double"
					},
					"item": {
						"$ref": "#/components/schemas/Item"
					},
					"itemId": {
						"type": "number",
						"format": "double"
					},
					"count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"cart",
					"cartId",
					"item",
					"itemId",
					"count"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICartPayload": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"finished": {
						"type": "boolean"
					},
					"cartItems": {
						"items": {
							"$ref": "#/components/schemas/CartItems"
						},
						"type": "array"
					}
				},
				"required": [
					"userId",
					"finished",
					"cartItems"
				],
				"type": "object",
				"additionalProperties": false
			},
			"File": {
				"description": "Provides information about files and allows JavaScript in a web page to access their content.",
				"properties": {
					"size": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"type": "string"
					},
					"lastModified": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"webkitRelativePath": {
						"type": "string"
					}
				},
				"required": [
					"size",
					"type",
					"lastModified",
					"name",
					"webkitRelativePath"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IItemPayload": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"categoryId": {
						"type": "number",
						"format": "double"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"promoId": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"$ref": "#/components/schemas/File"
					}
				},
				"required": [
					"name",
					"description",
					"categoryId",
					"price",
					"promoId",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPromoPayload": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"discount": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"$ref": "#/components/schemas/File"
					}
				},
				"required": [
					"name",
					"description",
					"discount",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISearchResult": {
				"properties": {
					"name": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"image": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"category",
					"id",
					"image"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICategoryPayload": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/File"
					}
				},
				"required": [
					"name",
					"description",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "express-typescript",
		"version": "1.1.1",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/users/all": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/User"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": []
			}
		},
		"/users": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserPayload"
							}
						}
					}
				}
			}
		},
		"/users/id/{id}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/User"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Boolean"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/exist/{email}": {
			"get": {
				"operationId": "Exist",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Boolean"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/count": {
			"get": {
				"operationId": "Count",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": []
			}
		},
		"/carts/all": {
			"get": {
				"operationId": "GetCarts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Cart"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Cart"
				],
				"security": [],
				"parameters": []
			}
		},
		"/carts": {
			"post": {
				"operationId": "CreateCart",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Cart"
								}
							}
						}
					}
				},
				"tags": [
					"Cart"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICartPayload"
							}
						}
					}
				}
			}
		},
		"/carts/all/{id}": {
			"get": {
				"operationId": "GetAllUserCarts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Cart"
									},
									"type": "array",
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Cart"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/carts/id/{id}": {
			"get": {
				"operationId": "GetCart",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Cart"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Cart"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteCart",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Boolean"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Cart"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/carts/current/{id}": {
			"get": {
				"operationId": "GetCurrentCart",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Cart"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Cart"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/items/all": {
			"get": {
				"operationId": "GetItems",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Item"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Item"
				],
				"security": [],
				"parameters": []
			}
		},
		"/items": {
			"post": {
				"operationId": "CreateItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Item"
								}
							}
						}
					}
				},
				"tags": [
					"Item"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IItemPayload"
							}
						}
					}
				}
			}
		},
		"/items/id/{id}": {
			"get": {
				"operationId": "GetItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Item"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Item"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Boolean"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Item"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/items/category/{categoryId}": {
			"get": {
				"operationId": "GetItemByCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Item"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Item"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/promos": {
			"get": {
				"operationId": "GetPromos",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Promo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Promo"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreatePromo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Promo"
								}
							}
						}
					}
				},
				"tags": [
					"Promo"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPromoPayload"
							}
						}
					}
				}
			}
		},
		"/promos/{id}": {
			"get": {
				"operationId": "GetPromo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Promo"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Promo"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeletePromo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Boolean"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Promo"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/search/{query}": {
			"get": {
				"operationId": "Search",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ISearchResult"
									},
									"type": "array",
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Search"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "query",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/categories": {
			"get": {
				"operationId": "GetCategorys",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Category"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Category"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICategoryPayload"
							}
						}
					}
				}
			}
		},
		"/categories/{id}": {
			"get": {
				"operationId": "GetCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Category"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Boolean"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}